+++ app-backend/common/src/test/scala/com/implicits/Generators.scala
@@ -8,7 +8,7 @@ import com.azavea.stac4s.Proprietary
 import eu.timepit.refined.auto._
 import eu.timepit.refined.types.string.NonEmptyString
 import geotrellis.vector.testkit.Rectangle
-import geotrellis.vector.{MultiPolygon, Point, Polygon, Projected}
+import geotrellis.vector.{io => _, _}
 import io.circe.syntax._
 import io.circe.testing.ArbitraryInstances
 import org.scalacheck.Arbitrary.arbitrary
@@ -74,20 +74,15 @@ object Generators extends ArbitraryInstances {
     AnnotationQuality.Unsure
   )
 
-  private def splitPeriodGen: Gen[SplitPeriod] = Gen.oneOf(
-    SplitPeriod.Day,
-    SplitPeriod.Week
-  )
-
   private def visibilityGen: Gen[Visibility] =
     Gen.oneOf(Visibility.Public, Visibility.Organization, Visibility.Private)
 
   private def taskStatusGen: Gen[TaskStatus] =
     Gen.frequency(
       (1, TaskStatus.Unlabeled),
-      (1, TaskStatus.LabelingInProgress),
+      (6, TaskStatus.LabelingInProgress),
       (1, TaskStatus.Labeled),
-      (1, TaskStatus.ValidationInProgress),
+      (6, TaskStatus.ValidationInProgress),
       (1, TaskStatus.Validated),
       (6, TaskStatus.Flagged),
       (1, TaskStatus.Invalid)
@@ -239,11 +234,14 @@ object Generators extends ArbitraryInstances {
       centerX <- Gen.choose(-2e7, 2e7)
       centerY <- Gen.choose(-2e7, 2e7)
     } yield {
-      Rectangle()
-        .withWidth(width)
-        .withHeight(height)
-        .setCenter(Point(centerX, centerY))
-        .build()
+      (Extent
+        .toPolygon(
+          (Rectangle()
+            .withWidth(width)
+            .withHeight(height)
+            .setCenter(Point(centerX, centerY))
+            .build(): Geometry).extent
+        ))
     }
 
   private def multiPolygonGen3857: Gen[MultiPolygon] =
@@ -891,39 +889,6 @@ object Generators extends ArbitraryInstances {
       )
     }
 
-  private def splitOptionsGen: Gen[SplitOptions] =
-    for {
-      name <- nonEmptyStringGen
-      colorGroupHex <- Gen.const(None)
-      t1 <- timestampIn2016Gen
-      t2 <- timestampIn2016Gen
-      period <- splitPeriodGen
-      onDatasource <- arbitrary[Option[Boolean]]
-      removeFromLayer <- arbitrary[Option[Boolean]]
-    } yield {
-      if (t1.before(t2)) {
-        SplitOptions(
-          name,
-          colorGroupHex,
-          t1,
-          t2,
-          period,
-          onDatasource,
-          removeFromLayer
-        )
-      } else {
-        SplitOptions(
-          name,
-          colorGroupHex,
-          t2,
-          t1,
-          period,
-          onDatasource,
-          removeFromLayer
-        )
-      }
-    }
-
   private def metricEventGen: Gen[MetricEvent] = Gen.oneOf(
     projectMosaicEventGen.widen,
     analysisEventGen.widen
@@ -973,7 +938,7 @@ object Generators extends ArbitraryInstances {
 
   private def taskPropertiesCreateGen: Gen[Task.TaskPropertiesCreate] =
     for {
-      status <- taskStatusGen
+      status <- Gen.const[TaskStatus](TaskStatus.Unlabeled)
       annotationProjectId <- uuidGen
       note <- if (status == TaskStatus.Flagged) {
         nonEmptyStringGen map { s =>
@@ -1335,10 +1300,6 @@ object Generators extends ArbitraryInstances {
     implicit def arbAnnotationQueryParameters
         : Arbitrary[AnnotationQueryParameters] = Arbitrary {
       annotationQueryParametersGen
-    }
-
-    implicit def arbSplitOptions: Arbitrary[SplitOptions] = Arbitrary {
-      splitOptionsGen
     }
 
     implicit def arbMetric: Arbitrary[Metric] = Arbitrary {