+++ app-backend/db/src/main/scala/ProjectLayerDao.scala
@@ -31,8 +31,10 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
 
   def deleteCache(id: UUID): ConnectionIO[Unit] = {
     for {
-      _ <- remove(ProjectLayer.cacheKey(id))(projectLayerCache,
-                                             async[ConnectionIO]).attempt
+      _ <- remove(ProjectLayer.cacheKey(id))(
+        projectLayerCache,
+        async[ConnectionIO]
+      ).attempt
     } yield ()
   }
 
@@ -47,14 +49,18 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
     selectAllColsF ++ fr"from" ++ tableF
 
   def getProjectLayerById(
-      projectLayerId: UUID): ConnectionIO[Option[ProjectLayer]] =
-    Cache.getOptionCache(ProjectLayer.cacheKey(projectLayerId),
-                         Some(30 minutes)) {
+      projectLayerId: UUID
+  ): ConnectionIO[Option[ProjectLayer]] =
+    Cache.getOptionCache(
+      ProjectLayer.cacheKey(projectLayerId),
+      Some(30 minutes)
+    ) {
       query.filter(projectLayerId).selectOption
     }
 
   def unsafeGetProjectLayerById(
-      projectLayerId: UUID): ConnectionIO[ProjectLayer] =
+      projectLayerId: UUID
+  ): ConnectionIO[ProjectLayer] =
     cachingF(ProjectLayer.cacheKey(projectLayerId))(Some(30 minutes)) {
       query.filter(projectLayerId).select
     }
@@ -64,17 +70,20 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
 
   def listProjectLayersForProject(
       page: PageRequest,
-      projectId: UUID): ConnectionIO[PaginatedResponse[ProjectLayer]] =
+      projectId: UUID
+  ): ConnectionIO[PaginatedResponse[ProjectLayer]] =
     listProjectLayersForProjectQ(projectId).page(page)
 
   def listProjectLayersWithImagery(
-      projectId: UUID): ConnectionIO[List[ProjectLayer]] = {
+      projectId: UUID
+  ): ConnectionIO[List[ProjectLayer]] = {
     val tableF =
       fr"project_layers left join scenes_to_layers on project_layers.id = scenes_to_layers.project_layer_id"
     val queryBuilder = Dao.QueryBuilder[ProjectLayer](
       selectAllColsF ++ fr"from" ++ tableF,
       tableF,
-      Nil)
+      Nil
+    )
     queryBuilder
       .filter(fr"scenes_to_layers.scene_id IS NOT NULL")
       .filter(fr"project_id = ${projectId}")
@@ -140,8 +149,9 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
   ): ConnectionIO[Option[ProjectLayer]] =
     query.filter(fr"project_id = ${projectId}").filter(layerId).selectOption
 
-  def deleteProjectLayer(layerId: UUID)(
-      implicit L: LiftIO[ConnectionIO]): ConnectionIO[Int] =
+  def deleteProjectLayer(
+      layerId: UUID
+  )(implicit L: LiftIO[ConnectionIO]): ConnectionIO[Int] =
     for {
       pl <- unsafeGetProjectLayerById(layerId)
       _ <- pl.overviewsLocation match {
@@ -160,85 +170,15 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
     } yield updateQuery
   }
 
-  def batchCreateLayers(
-      groupedScenes: Map[(Option[(Timestamp, Timestamp)], Option[String]),
-                         List[Scene.ProjectScene]],
-      layer: ProjectLayer,
-      splitOptions: SplitOptions): ConnectionIO[List[ProjectLayer]] = {
-    val projectLayersAndScenes
-      : Map[ProjectLayer.Create, List[Scene.ProjectScene]] = groupedScenes.map {
-      case ((Some((start, end)), datasourceO), scenes) =>
-        (ProjectLayer.Create(
-           datasourceO match {
-             case Some(datasource) =>
-               s"${splitOptions.name} | " +
-                 s"${datasource}"
-             case _ =>
-               s"${splitOptions.name}"
-           },
-           layer.projectId,
-           splitOptions.colorGroupHex.getOrElse(layer.colorGroupHex),
-           Some(layer.id),
-           Some(start),
-           Some(end),
-           layer.geometry,
-           layer.isSingleBand,
-           layer.singleBandOptions,
-           layer.overviewsLocation,
-           layer.minZoomLevel
-         ),
-         scenes)
-      case ((_, datasourceO), scenes) =>
-        (ProjectLayer.Create(
-           datasourceO match {
-             case Some(datasource) =>
-               s"${splitOptions.name} | " +
-                 s"${datasource}"
-             case _ =>
-               s"${splitOptions.name}"
-           },
-           layer.projectId,
-           splitOptions.colorGroupHex.getOrElse(layer.colorGroupHex),
-           Some(layer.id),
-           None,
-           None,
-           layer.geometry,
-           layer.isSingleBand,
-           layer.singleBandOptions,
-           layer.overviewsLocation,
-           layer.minZoomLevel
-         ),
-         scenes)
-    }
-
-    projectLayersAndScenes.toList traverse {
-      case (projectLayerC: ProjectLayer.Create,
-            scenes: List[Scene.ProjectScene]) =>
-        (scenes.toNel, layer.projectId) match {
-          case (Some(s), Some(pId)) =>
-            for {
-              insertedLayer <- insertProjectLayer(projectLayerC.toProjectLayer)
-              _ <- ProjectDao.addScenesToProject(
-                s.map(_.id),
-                pId,
-                insertedLayer.id,
-                true
-              )
-            } yield insertedLayer
-          case _ =>
-            throw new java.lang.IllegalArgumentException(
-              s"Cannot add scenes to a layer which is not in a project: ${layer.id}"
-            )
-        }
-    }
-  }
-
   def getDayRangeFromTimestamp(date: Timestamp): (Timestamp, Timestamp) = {
     val startOfDay =
       date.toLocalDateTime.toLocalDate.atStartOfDay
-    (new Timestamp(startOfDay.toEpochSecond(ZoneOffset.UTC) * 1000),
-     new Timestamp(
-       startOfDay.plusHours(24).toEpochSecond(ZoneOffset.UTC) * 1000))
+    (
+      new Timestamp(startOfDay.toEpochSecond(ZoneOffset.UTC) * 1000),
+      new Timestamp(
+        startOfDay.plusHours(24).toEpochSecond(ZoneOffset.UTC) * 1000
+      )
+    )
   }
 
   def getWeekRangeFromTimestamp(date: Timestamp): (Timestamp, Timestamp) = {
@@ -251,53 +191,16 @@ object ProjectLayerDao extends Dao[ProjectLayer] {
       .`with`(TemporalAdjusters.previousOrSame(DayOfWeek.MONDAY))
 
     val startOfDay = datetimeWeek.atStartOfDay
-    (new Timestamp(startOfDay.toEpochSecond(ZoneOffset.UTC) * 1000),
-     new Timestamp(startOfDay.plusDays(7).toEpochSecond(ZoneOffset.UTC) * 1000))
+    (
+      new Timestamp(startOfDay.toEpochSecond(ZoneOffset.UTC) * 1000),
+      new Timestamp(startOfDay.plusDays(7).toEpochSecond(ZoneOffset.UTC) * 1000)
+    )
   }
 
-  def groupScenesBySplitOptions(splitOptions: SplitOptions)
-    : Scene.ProjectScene => (Option[(Timestamp, Timestamp)], Option[String]) = {
-    scene: Scene.ProjectScene =>
-      (splitOptions.period, splitOptions.splitOnDatasource) match {
-        case (SplitPeriod.Day, Some(true)) =>
-          (scene.filterFields.acquisitionDate.map(getDayRangeFromTimestamp),
-           Some(scene.datasource.name))
-        case (SplitPeriod.Week, Some(true)) =>
-          (scene.filterFields.acquisitionDate.map(getWeekRangeFromTimestamp),
-           Some(scene.datasource.name))
-        case (SplitPeriod.Day, _) =>
-          (scene.filterFields.acquisitionDate.map(getDayRangeFromTimestamp),
-           None)
-        case (SplitPeriod.Week, _) =>
-          (scene.filterFields.acquisitionDate.map(getWeekRangeFromTimestamp),
-           None)
-      }
-  }
-
-  def splitProjectLayer(
-      projectId: UUID,
+  def layerIsInProject(
       layerId: UUID,
-      splitOptions: SplitOptions): ConnectionIO[List[ProjectLayer]] = {
-    for {
-      layer <- unsafeGetProjectLayerById(layerId)
-      scenes <- ProjectLayerScenesDao
-        .listLayerScenesRaw(layerId, Some(splitOptions))
-        .flatMap(ProjectLayerScenesDao.scenesToProjectScenes(_, layerId))
-      groupedScenes = scenes.groupBy(groupScenesBySplitOptions(splitOptions))
-      newLayers <- batchCreateLayers(groupedScenes, layer, splitOptions)
-      _ <- deleteCache(layerId)
-      _ <- splitOptions.removeFromLayer match {
-        case Some(true) =>
-          ProjectDao.deleteScenesFromProject(scenes.map(_.id),
-                                             projectId,
-                                             layerId)
-        case _ => 0.pure[ConnectionIO]
-      }
-    } yield newLayers
-  }
-
-  def layerIsInProject(layerId: UUID,
-                       projectID: UUID): ConnectionIO[Boolean] = {
+      projectID: UUID
+  ): ConnectionIO[Boolean] = {
     getProjectLayerById(layerId) map {
       case Some(projectLayer) => projectLayer.projectId == Option(projectID)
       case _                  => false